Ninth Push, All Phase 1 Clear

+++ database/migrations/2024_07_08_043009_create_bbins_table.php
@@ -26,11 +26,11 @@ return new class extends Migration
             $table->string('item_longdescription')->nullable();
             $table->string('item_uofm');
             $table->integer('total_container');
-            $table->decimal('total_quantity', 12, 4);
-            $table->decimal('quantity_remaining', 12, 4);
+            $table->double('total_quantity');
+            $table->double('quantity_remaining');
             $table->unsignedBigInteger('currency_id');
             $table->foreign('currency_id')->references('id')->on('currencies');
-            $table->decimal('item_amount', 12, 4);
+            $table->double('item_amount');
             $table->unsignedBigInteger('storages_id');
             $table->foreign('storages_id')->references('id')->on('storages');
             $table->unsignedBigInteger('subkontrak_id')->nullable();
@@ -39,7 +39,7 @@ return new class extends Migration
             $table->foreign('supplier_id')->references('id')->on('suppliers');
             $table->unsignedBigInteger('country_id');
             $table->foreign('country_id')->references('id')->on('countries');
-            $table->decimal('kurs', 12, 4);
+            $table->double('kurs');
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users');
             $table->string('user_name');
